@@ -10,7 +10,7 @@
     }
   ],
   "require": {
-    "php": ">=7.1"
+    "php": ">=8.2"
   },
   "autoload": {
     "psr-4": {
